Add first implementation of collecting failed sources
need to integrate the sdk to really test it but have the groundwork to
move quickly on this. just need to adjust one call
this script runs through an array of geonode configs
it picks all the sources and tries to create them with the boundless sdk
if that fails (sdk does not return an array of layers) it failed and we
need to file a report. save the source types in an array for now. but
later add more information to it.